chore: splitted pixel and script
