+++ pages/_document.tsx
@@ -1,25 +1,8 @@
+import Pixel from "@/components/Pixel"
 import { Html, Head, Main, NextScript } from "next/document"
 import Script from "next/script"
-import { useRouter } from "next/router"
-import * as fbq from '../lib/fb-pixel'
-import { useEffect } from "react"
 
 export default function Document() {
-  const router = useRouter()
-
-  useEffect(() => {
-    // This pageview only triggers the first time (it's important for Pixel to have real information)
-    // fbq.pageview()
-
-    const handleRouteChange = () => {
-      fbq.pageview()
-    }
-
-    router.events.on('routeChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
   return (
     <Html lang="es">
       <Head>
@@ -34,24 +17,6 @@ export default function Document() {
           />
         </noscript>
       </Head>
-      <Script
-        id="fb-pixel"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            !function(f,b,e,v,n,t,s)
-            {if(f.fbq)return;n=f.fbq=function(){n.callMethod?
-            n.callMethod.apply(n,arguments):n.queue.push(arguments)};
-            if(!f._fbq)f._fbq=n;n.push=n;n.loaded=!0;n.version='2.0';
-            n.queue=[];t=b.createElement(e);t.async=!0;
-            t.src=v;s=b.getElementsByTagName(e)[0];
-            s.parentNode.insertBefore(t,s)}(window, document,'script',
-            'https://connect.facebook.net/en_US/fbevents.js');
-            fbq('init', 487461332128996);
-            fbq('track', 'PageView');
-          `,
-        }}
-      />
       <body>
 
         <Main />
